fixed warning on x64

+++ src/aho_corasick/aho_corasick.hpp
@@ -116,11 +116,11 @@ namespace aho_corasick {
 			}
 
 			size_t determine_median(const interval_collection& intervals) const {
-				int start = -1;
-				int end = -1;
+				size_t start = -1;
+				size_t end = -1;
 				for (const auto& i : intervals) {
-					int cur_start = i.get_start();
-					int cur_end = i.get_end();
+					size_t cur_start = i.get_start();
+					size_t cur_end = i.get_end();
 					if (start == -1 || cur_start < start) {
 						start = cur_start;
 					}
